requirements.txt
pandas
numpy
ccxt
requests
python-dotenv
matplotlib